+++ changelog.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 # [Unreleased]
 
-## [0.2.0] - 2026-08-27
+## [0.2.1] - 2026-08-27
 
 ## Fixed
 
-- Doc only: Correct doc comment about how elbow anchor choice is evaluated (``)
+- Doc only: Correct doc comment about how elbow anchor choice is evaluated (`9b93829`)
+- Remove repeated `Vec<Point>` allocation during rerouting (``)
 
 # [Released]